pyproject.toml
[tool.poetry]
name = "bioimage_analysis_framework_demo"
version = "0.1.0"
description = "Demo application for the position Python programmer at Danish BioImaging Infrastructure Image Analysis Core Facility (DBI-INFRA IACF)"
authors = ["Jonas Manthey <jonas@amps.one>"]


[tool.poetry.dependencies]
python = ">=3.10,<4.0"
loguru = "^0.6.0"
celery = "^5.2.7"
redis = "^4.4.0"

[tool.poetry.extras]


[tool.poetry.dev-dependencies]
autodoc-pydantic = "^1.7.2"
black = "^22.3.0"
celery-types = "^0.13.1"
flake8 = "^5.0.4"
flake8-comprehensions = "^3.10.0"
flake8-eradicate = "^1.2.1"
flake8-expression-complexity = "^0.0.11"
flake8-functions = "^0.0.7"
flake8-pep585 = "^0.1.5"
flake8-pep604 = "^0.1.0"
flake8-use-fstring = "^1.3"
isort = "^5.10.1"
mypy = "^0.971"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
exclude = '''
/(
    \.git
  | \.mypy_cache
  | \.tox
  | venv
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black" 